Resolve bugs tied to record creations on table
