+++ packages/twenty-front/src/modules/object-record/record-index/states/recordIndexViewTypeState.ts
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+import { ViewType } from '@/views/types/ViewType';
+
+export const recordIndexViewTypeState = atom<ViewType | undefined>({
+  key: 'recordIndexViewTypeState',
+  default: undefined,
+});
