+++ packages/twenty-front/src/pages/object-record/RecordIndexPage.tsx
@@ -1,19 +1,15 @@
 import { useParams } from 'react-router-dom';
 import styled from '@emotion/styled';
 
-import { useObjectMetadataItemForSettings } from '@/object-metadata/hooks/useObjectMetadataItemForSettings';
 import { useObjectNameSingularFromPlural } from '@/object-metadata/hooks/useObjectNameSingularFromPlural';
 import { useCreateOneRecord } from '@/object-record/hooks/useCreateOneRecord';
 import { RecordIndexContainer } from '@/object-record/record-index/components/RecordIndexContainer';
 import { DEFAULT_CELL_SCOPE } from '@/object-record/record-table/record-table-cell/hooks/useOpenRecordTableCell';
 import { useSelectedTableCellEditMode } from '@/object-record/record-table/record-table-cell/hooks/useSelectedTableCellEditMode';
-import { useIcons } from '@/ui/display/icon/hooks/useIcons';
-import { PageAddButton } from '@/ui/layout/page/PageAddButton';
 import { PageBody } from '@/ui/layout/page/PageBody';
 import { PageContainer } from '@/ui/layout/page/PageContainer';
-import { PageHeader } from '@/ui/layout/page/PageHeader';
-import { PageHotkeysEffect } from '@/ui/layout/page/PageHotkeysEffect';
 import { useSetHotkeyScope } from '@/ui/utilities/hotkey/hooks/useSetHotkeyScope';
+import { RecordIndexPageHeader } from '~/pages/object-record/RecordIndexPageHeader';
 
 const StyledIndexContainer = styled.div`
   display: flex;
@@ -28,20 +24,11 @@ export const RecordIndexPage = () => {
     objectNamePlural,
   });
 
-  const { findObjectMetadataItemByNamePlural } =
-    useObjectMetadataItemForSettings();
-
-  const { getIcon } = useIcons();
-  const Icon = getIcon(
-    findObjectMetadataItemByNamePlural(objectNamePlural)?.icon,
-  );
-
   const { createOneRecord: createOneObject } = useCreateOneRecord({
     objectNameSingular,
   });
 
   const recordIndexId = objectNamePlural ?? '';
-
   const setHotkeyScope = useSetHotkeyScope();
 
   const { setSelectedTableCellEditMode } = useSelectedTableCellEditMode({
@@ -49,7 +36,9 @@ export const RecordIndexPage = () => {
   });
 
   const handleAddButtonClick = async () => {
-    await createOneObject?.({});
+    await createOneObject?.({
+      position: 0,
+    });
 
     setSelectedTableCellEditMode(0, 0);
     setHotkeyScope(DEFAULT_CELL_SCOPE.scope, DEFAULT_CELL_SCOPE.customScopes);
@@ -57,15 +46,7 @@ export const RecordIndexPage = () => {
 
   return (
     <PageContainer>
-      <PageHeader
-        title={
-          objectNamePlural.charAt(0).toUpperCase() + objectNamePlural.slice(1)
-        }
-        Icon={Icon}
-      >
-        <PageHotkeysEffect onAddButtonClick={handleAddButtonClick} />
-        <PageAddButton onClick={handleAddButtonClick} />
-      </PageHeader>
+      <RecordIndexPageHeader createRecord={handleAddButtonClick} />
       <PageBody>
         <StyledIndexContainer>
           <RecordIndexContainer
